caching: rug histogram
